Fixed a bug that comes with number of events with neutrons, results in extremely big number for arrival time. Now skipping those hits where arrival time is expectedly high.

+++ PSD/include/PSD_PMT_Hit.h
@@ -3,13 +3,13 @@
 
 #include "G4VHit.hh"
 class PSD_PMT_Hit : public G4VHit {
-  int fPhotonArrivalTime;
+  long unsigned int fPhotonArrivalTime;
 
 public:
   PSD_PMT_Hit();
   virtual ~PSD_PMT_Hit();
-  void SetArrivalTime(int tme);
-  int GetArrivalTime() const;
+  void SetArrivalTime(long unsigned int tme);
+  long unsigned int GetArrivalTime() const;
   void Print();
 };
